requirements.txt
EbookLib==0.18
lxml==4.9.2
Pillow==9.5.0
six==1.16.0
